backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
httpx
pyjwt
openai
python-dotenv
